vglconnect: Disable X11 fwding w/ nettest SSH cmd
Referring to https://bugzilla.redhat.com/show_bug.cgi?id=1972266, if X11
forwarding is enabled by default in the SSH client (for instance, by
setting 'X11Forward=yes' in ~/.ssh/config) and the SSH server (for
instance, by setting 'X11Forwarding yes' in /etc/ssh/sshd_config), then
attempting to execute a non-interactive command remotely via SSH on a
RHEL/CentOS 8 server exposes a bug whereby control is never returned to
the console, effectively causing the command to hang.  The easiest
workaround is to explicitly pass -x to ssh in vglconnect when finding a
free port on the server, since X11 forwarding is unnecessary for that
particular command.

Closes #169