EGLBE: Fix segfault w/piglit glx-fbconfig-bad test
glx-fbconfig-bad abuses the internals of GLX by passing the address of a
__GLXFBConfig struct, with all members set to 0, to
glXCreateNewContext().  VirtualGL 3.0 interprets GLXFBConfig as a
pointer to a _VGLFBConfig struct, so effectively the glx-bad-fbconfig
shenanigans cause the rboCtx structure member to be NULL.  The EGL back
end did not previously account for this scenario, since it could never
occur with a real 3D application.  glx-fbconfig-bad now triggers a
GLXBadConfig error when used with the EGL back end, as expected.