EGLBE: Don't print 3D X server true color warning
That warning message is a throwback to the Sun Microsystems days when
some users were still running X servers with 8-bit or 16-bit color
depths.  Since VirtualGL, when using the GLX back end, requires a 24-bit
color depth on both the 2D X server and 3D X server, the warning message
was an attempt to encourage users to RTFM regarding that requirement
rather than seek tech support when their 3D applications, which worked
fine on the local display, failed to find visuals when using VirtualGL.
The warning message is not relevant to the EGL back end.

Refer to #183