EGL/X11: Opt in rather than out of EGL extensions
It is safer to explicitly claim support for EGL extensions that will
probably work rather than to explicitly disclaim support for EGL
extensions that definitely won't work.  If we overlook an extension with
the former approach, then the overall functionality will be artificially
reduced, but if we overlook an extension with the latter approach, then
an application crash or other unexpected behavior could occur.