Build: Don't cache failed TurboJPEG test status
CMake caches the results of the TURBOJPEG_WORKS and
SYSTEM_TURBOJPEG_WORKS tests.  If either of those tests fails,
presumably because TJPEG_LIBRARY is incorrect, then we need to uncache
both test status variables to ensure that CMake re-runs the tests if it
is re-invoked with a different TJPEG_LIBRARY value.

Fixes #188