Doc: w/ recent GDM, gdm user must be in vglusers
... if framebuffer device access has been restricted to vglusers.

Refer to https://github.com/VirtualGL/virtualgl/issues/65#issuecomment-1076296650.

In my testing, the GDM implementations in Ubuntu 18.04+ and Fedora (but
not the ones in RHEL/CentOS, as of this writing) run the greeter X
server (which, from VGL's point of view, is the same as the 3D X server)
under the gdm account rather than the root account, so gdm must be a
member of vglusers if framebuffer device access has been restricted to
vglusers.  Otherwise, the 3D X server will fail to start.