Faker: Interpose XCB window creation/destr. funcs
We need to track X windows created using the XCB API.  Otherwise, if one
of those windows is used with GLX, the faker erroneously assumes that
the window was created in another process, which causes the faker to
create a VirtualWin instance that is not destroyed until the display
connection is closed or the application exits.  Since VirtualGL
maintains as many as four X display connections for each VirtualWin
instance, failing to destroy VirtualWin instances in a timely manner can
lead to X resource exhaustion if an application attempts to create and
destroy hundreds of X windows using the same X display connection.

Fixes #195