glXChooseFBConfig: Only return configs in CA table
For reasons unexplained, the internal build of Mesa that MATLAB uses
when it is invoked with -softwareopengl returns an FB config in
glXChooseFBConfig() (specifically FB config 0x23) that it does not
return in glXGetFBConfigs().  When using the GLX back end, VirtualGL's
interposed glXChooseFBConfig() function assumed that the aforementioned
situation could never happen.  Thus, when it failed to find a match in
the config attribute table for the first (and only) FB config returned
by the real glXChooseFBConfig() function, the interposed
glXChooseFBConfig() function returned a non-zero element count along
with a GLXFBConfig array that had a NULL first entry.  A segfault
occurred when the calling application subsequently tried to de-reference
the first entry.

This fix is undocumented, since running MATLAB with -softwareopengl and
VirtualGL is neither a supported nor a useful configuration.

Based on:
https://github.com/nathankidd/virtualgl/commit/a18e4185162b7cad42c0987efe3b9e45d301a94a

Closes #201