EGLBE: Allow specifying an EGL device without DRI
EGL doesn't actually require DRI, so it isn't necessary to use a DRI
device path to specify an EGL device.  The EGL device can simply be
specified by an "EGL device ID" of the form egl{n}, where {n} is the
device's index in the array of valid EGL devices returned by
eglQueryDevicesEXT().  This facilitates using the EGL back end on
systems that, for whatever reason, have not loaded the nvidia_drm
module.  This also allows the Mesa unaccelerated EGL device to be
specified, although the usefulness of that is doubtful.

Closes #207