EGLBE: Restore buf state when binding default FB
This is very subtle, but glBindFramebuffer(..., 0) should restore the
current context's previous draw/read buffer state for the default
framebuffer, as glXMake*Current() does.  Swapping the buffers of a GLX
drawable causes the EGL back end to create a new FBO to serve as the
drawable's default framebuffer, with the new FBO containing the previous
RBOs bound in reverse order.  The default draw/read buffer for a new FBO
is GL_COLOR_ATTACHMENT0, which the EGL back end uses as the front left
buffer.  Thus, if the default framebuffer for a drawable isn't bound
when the drawable's buffers are swapped, then the FBO's initial
draw/read buffer remains unchanged.  If GL_BACK was active the last time
the default framebuffer was bound, then calling
glBindFrameBuffer(..., 0) after a buffer swap put VirtualGL into a weird
state whereby it reported that GL_BACK was active, but the RBO
corresponding to GL_FRONT was actually active.  This issue caused Webots
to flicker when rendering a camera view, and it may have affected other
applications as well.

Fixes #172