EGLBE: Incr. RBO ctx ref. when creating shared ctx
If a context is shared with another context, then it is indirectly
shared with the RBO context, so in the body of glXCreate*Context*()
(i.e. in the body of backend::createContext()), we should always call
RBOContext::createContext() regardless of whether 'share' is NULL.  If
'share' is not NULL, then the RBO context will have already been created
in a previous call to glXCreate*Context*(), so calling
RBOContext::createContext() again will increment the context reference
count for the RBO context.  That is necessary because (a) the RBO
context won't truly be unused until all fake Pbuffers and all contexts
that are shared with it (either directly or indirectly) are destroyed,
and (b) the reference count is always decremented in the body of
glXDestroyContext(), regardless of whether the context is shared.

This fixes GL_INVALID_OPERATION or GL_FRAMEBUFFER_INCOMPLETE_ATTACHMENT
errors (regression caused by 2e94a75f82a1f99b615b68a0d4e3f616cebb9527)
that occurred in FakePbuffer::createBuffer() on certain systems when a
Qt OpenGL window was destroyed and another was created.