EGLBE: Store current GLX ctx in thread-local var
... instead of passing through the current EGL context.

The main purpose of this is to fix an esoteric use case whereby a
particular 3D application (Gazebo), in the process of probing GLX and
EGL capabilities,
- creates a GLX context and makes it current,
- creates a device-based EGL context bound to the desktop OpenGL API and
  makes it current, and
- expects the glXGetCurrent*() functions to return values as if the GLX
  context is current, even though it actually isn't.

I would (and did) argue that, even though this use case works with
nVidia's proprietary driver and Mesa-based drivers (including AMDGPU),
those drivers are behaving incorrectly, and Gazebo is relying on the
incorrect behavior.  GLX and EGL contexts are opaque wrapper objects for
OpenGL contexts, and two OpenGL contexts cannot simultaneously be
current in the same thread.  If the glXGetCurrent*() functions return
non-zero values, then the calling program could assume that it can:

- import X11 fence objects, via the GL_EXT_x11_sync_object extension,
  from the display handle returned by glXGetCurrentDisplay() into the
  current OpenGL context,
- call glXWaitGL() and glXWaitX() to synchronize X11 and OpenGL
  rendering, and
- use X11 functions (XGetImage(), etc.), with the drawable(s) returned
  by glXGetCurrent*Drawable(), to obtain or modify the OpenGL-rendered
  pixels.

Of course, all of those assumptions are false.

However, the behavior of VGL 3.0.x was also not entirely correct,
because glXGetCurrentContext() returned the device-based EGL context
rather than the GLX context.  IMHO, the actual correct behavior would be
for the glXGetCurrent*() functions to return 0 whenever an EGL context
is current.  However, because the nVidia and AMD/Mesa implementations
don't do that, implementing that behavior in the EGL back end would
effectively cause its behavior to differ from that of the GLX back end.
Also, it would be impossible to implement that behavior without an EGL
interposer (front end.)