Faker: Don't allow EGLBE display to be terminated
eglGetPlatformDisplay*() returns a static EGLDisplay handle for each EGL
device.  Thus, we cannot allow the 3D application to call eglTerminate()
with the EGLDisplay handle used by the EGL back end, or hilarity will
ensue.  As with the previous commit, this is intended to allow
applications to use GLX and device-based EGL simultaneously, even though
that is not a good idea.