EGLX: Fail gracefully if EGL back end not in use
Some applications, including Firefox, support both EGL/X11 and GLX, so
VGL shouldn't throw a fatal error if those applications call
eglGet*Display() to probe for EGL/X11 support.

I suspect that this fixes the Firefox complaint in #228.

NOTE: From a usability point of view, it would seemingly be nicer to
automatically use an EGL device with the EGL/X11 interposer, but that
would require introducing a separate environment variable/command-line
option so that an EGL device could be specified for the EGL/X11
interposer without forcing the EGL back end to be used with the GLX
interposer.  At the moment, it makes more sense to tie the EGL/X11
interposer to the EGL back end, since the GLX back end is primarily a
backward compatibility feature.