Faker: Ignore bits 9- in XKeyEvent state field
XKB uses bits 13-14 of the XKeyEvent state field to store the group
index.  If the X keyboard extension was enabled on the 2D X server
(which is always the case with modern X servers), that prevented
handleEvent() (and presumably handleXCBEvent() as well) from ever
matching a KeyPress event with the key sequence used to pop up the
VirtualGL Configuration dialog.  Thus, the dialog never popped up.  All
of the modifier key masks use bits 1-8, so we can safely ignore any
other bits.