Faker: Automatically set VGL_PROBEGLX if unset
Probing the 2D X server for stereo visuals is useless unless the VGL
transport or a transport plugin will be used.  Normally VGL doesn't
determine that until a VirtualWin instance is created, which is well
after the visual attribute table is built.  However, we have enough
information at the time the visual attribute table is built to determine
which transport will be used.  If the X11 transport will be used
initially, then it will not be possible to switch to another transport
using the VirtualGL Configuration dialog, so it is safe to set
VGL_PROBEGLX to 0 if it has not been set explicitly.

Probing the 2D X server for stereo visuals can cause problems if the
X client's OpenGL implementation doesn't play nicely with the 2D X
server's OpenGL implementation or if the 3D application includes a
static build of Mesa.  (Referring to
https://groups.google.com/g/virtualgl-users/c/Gy0Y0OmrEpU/m/RdG1PewaAgAJ
this is a known issue with Tecplot 360.)  TurboVNC 3.0 and later
automatically sets VGL_PROBEGLX=0, but that only affects applications
launched within the TurboVNC session.