Eliminate usage of the 'register' keyword
The 'register' keyword was deprecated in C++11 and removed in C++17, so
usage of that keyword in VirtualGL (more specifically in the in-tree
FLTK and MD5 code, not in the VGL code itself) prevented successful
compilation with newer compilers (e.g. Clang 16) unless -Wno-register
was specified.

Fixes https://bugs.gentoo.org/898876
Closes #233