EGLX/eglSwapBuffers(): Flush context before swap
The documentation for eglSwapBuffers() states that "eglSwapBuffers
performs an implicit flush operation on the context ... bound to surface
before swapping."  The GLX front end uses double-buffered Pbuffers to
emulate double-buffered windows, so it calls the real glXSwapBuffers()
function to swap the buffers, and the real glXSwapBuffers() function
ensures that OpenGL rendering has completed.  However, EGL Pbuffer
surfaces are always single-buffered, so it would be useless for the
EGL/X11 front end to call the real eglSwapBuffers() function.  Thus, we
have to emulate it.  glFlush() is asynchronous whereas glFinish() is
synchronous, so we need to call glFinish() if the current surface is
being swapped.  This fixes a flickering issue that occurred with recent
versions of Firefox.