Build: Always require EGL
All supported platforms now include sufficiently new EGL libraries to
build both the EGL back end and the EGL/X11 front end, and the ability
to build VirtualGL without those features is not well-tested and is
tricky to maintain.  The ability to build VirtualGL without EGL will
continue to be maintained in the 3.0.x branch on a break/fix basis.

Note that this eliminates the ability to build VirtualGL on Solaris.
However, we have not officially supported Solaris for quite some time
(https://virtualgl.org/Documentation/OSSupport), and the only reason why
I have even bothered to test VirtualGL on Solaris is because nVidia
continues to release drivers for it.  Since the VirtualGL 3.0.x branch
is still in the Maintenance support category, all non-disruptive fixes
have been back-ported to it from the main (3.1.x) branch (and will
continue to be back-ported until VirtualGL 3.2 is released), so the only
functional differences between the tip of 3.0.x and the tip of main
relate to EGL features that aren't relevant for Solaris.