Doc: Update Chrome app recipe
Referring to #229, passing --use-gl=egl to Chrome v112+ causes the GPU
process to crash, even when Chrome is run on a local display without
VirtualGL.  At the moment, there is no clean way to use Chrome/ANGLE
with VirtualGL, because when using ANGLE, Chrome obtains visuals using
raw GLX requests/replies (so there is nothing for VirtualGL to
interpose.)  Through some really messy modifications, VirtualGL can be
made to work with Chrome/ANGLE, but the performance is awful.  From
VirtualGL's point of view, these are pathological behaviors that the
Chrome developers need to resolve, much as the GTK developers resolved
them.  Until then, we cannot claim that Chrome works with VGL.