dlfaker: Interpose "libEGL.so", not "*/libEGL.so"
Referring to 211f71937d4f88363eef3e459f4f60479ba0d26c, it turns out that
Firefox is an example of a 3D application that calls
dlopen("libEGL.so", ...).  (This is why we can't have nice things.)
However, when ANGLE opens its EGL interposer, it always does so using a
full pathname.  Thus, it is (hopefully!) safe to assume that VirtualGL's
dlopen() interposer should interpose calls to dlopen("libEGL.so", ...)
but not calls to dlopen("*/libEGL.so", ...).