Faker: Emit correct error on exit when running without GLX
Commit bdc83ac67d7df6c7c62a3ea96af7ad0d730f5f92 introduced emitting GLX
error instead of throwing exceptions. However, when GLX is not enabled,
the faker will exit writing "The EGL back end requires a 2D X server
with a GLX extension." to stderr on any error. This hides the original
error and is not a valid way to indicate what failed.

This fix emits error details instead of erroneously reporting that GLX
should be enabled.