EGLBE: Free EGL resources when no context bound
When the last OpenGL context is unbound from a rendering thread (i.e.
when EGL_NO_CONTEXT is explicitly bound to the thread), we should call
eglReleaseThread() to free any GPU and API resources previously
allocated for the thread.  The most obvious issue this addresses is an
issue whereby, when using nVidia's proprietary drivers, running a
non-OpenGL application with VirtualGL caused the application to show up
in nvidia-smi (with non-zero GPU memory usage in some cases.)  This is
because of code in buildCfgAttribTable() that creates and binds a
temporary EGL context in order to probe the value of GL_MAX_SAMPLES,
which is necessary in order to determine which OpenGL framebuffer
configurations can successfully be emulated using renderbuffer objects.

Fixes #272