Faker: Improve err msg if XOpenDisplay !interposed
If a 3D application uses dlopen()/dlsym() to load X11 functions, then
libdlfaker.so must be preloaded.  Otherwise VirtualGL's interposed
version of XOpenDisplay() will not be called, the X11 display's
excluded status will not be added as a display extension, and
faker::isDisplayExcluded() and faker::getDisplayCS() will fail.  This
commit adds a more instructive error message if one of those methods
fails.

Refer to #214