Server: Don't disable Wayland; Use Display :1024
On modern Linux distributions, GDM runs Xwayland on Display :1024 if
WaylandEnable=true (the default), and it runs Xorg on Display :0 if
WaylandEnable=false.  Referring to
aa5ff87b9a7a569a62ffe3825265b447b863b57f and #94, in 2019 I observed
that it was necessary on RHEL 8 and Fedora 27-28 to add/uncomment
WaylandEnable=false in order for VirtualGL's greeter autostart script
to work.  However, that is apparently no longer necessary.  In 2025,
using Display :1024 seems to work fine on RHEL 7-10, Ubuntu 16-24,
Fedora 40-42, and openSUSE 15.  (In fact, it is necessary on RHEL 10,
since GDM no longer allows Wayland to be disabled.  That is what
prompted this change.)